requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.5.2
crewai==0.36.0
crewai-tools==0.10.1
pydantic==2.6.1
requests==2.31.0
serper==1.1.2
serperdev==0.2.1
gunicorn==21.2.0
